pyproject.toml
[project]
name = "educational-service-test"
version = "0.1.0"
description = ""
requires-python = ">=3.13"
dependencies = [
    "dj-database-url>=2.3.0",
    "django>=5.1.7",
    "django-rest-framework>=0.1.0",
    "dotenv>=0.9.9",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-decouple>=3.8",
]
